backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic[email]==2.10.4
pydantic-settings==2.7.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
httpx==0.28.1
python-multipart==0.0.20





google-auth==2.37.0
requests==2.32.3
pytest==8.3.4
pytest-asyncio==0.25.0
google-genai
Pillow
openai
